requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.21.1
packaging==24.2
pymongo==4.11.3
python-dotenv==1.1.0
Werkzeug==3.1.3
